[FIX] login 뷰 & 로직 수정
